+++ quiz/app/src/main/java/com/example/quiz/FewAnswersFiveOptions.java
@@ -35,7 +35,7 @@ public class FewAnswersFiveOptions extends Fragment {
 
         //get quiz id from global variable
         singleToneClass singleToneClass = com.example.quiz.singleToneClass.getInstance();
-        int quiz_id = singleToneClass.getData();
+        long quiz_id = singleToneClass.getData();
 
         Quiz quiz1 = db.getQuiz(quiz_id);
         String qn = quiz1.getQuizName();
